Move specialization printf_arument_type_traits for simd_view to esimd header
